Made it so doctors can add users to database. Also moved some css and javascript to appropriate files
