+++ MedicalSuiteWeb/Pages/Shared/_LayoutDoctor.cshtml
@@ -63,6 +63,9 @@
                                 <a class="dropdown-item" asp-area="" asp-page="/Users/ViewUsers">
                                     <i class="bi bi-people"></i> View Users
                                 </a>
+                                <a class="dropdown-item" asp-area="" asp-page="/Users/AddUser">
+                                    <i class="bi bi-person-add"></i> Add User
+                                </a>
                             </div>
                         </li>
