+++ MedicalSuiteWeb/Pages/Shared/_LayoutNurse.cshtml
@@ -50,6 +50,16 @@
                                 <a class="dropdown-item" asp-area="" asp-page="/Inventories/ViewInventoryItems">View Inventory</a>
                             </div>
                         </li>
+                        <li class="nav-item dropdown">
+                            <a class="nav-link dropdown-toggle text-dark" href="#" id="navbarDropdownMenuLink3" role="button" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                                Users
+                            </a>
+                            <div class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink3">
+                                <a class="dropdown-item" asp-area="" asp-page="/Users/ViewUsers">
+                                    <i class="bi bi-people"></i> View Users
+                                </a>
+                            </div>
+                        </li>
                         <!--
                         <li class="nav-item dropdown">
                             <a class="nav-link dropdown-toggle text-dark" href="#" id="navbarDropdownMenuLink3" role="button" data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
