+++ MedicalSuiteWeb/Pages/Users/ViewUsers.cshtml
@@ -22,7 +22,7 @@
                 <label class="form-label" asp-for="@user.FirstName"></label>
             </div>
             <div class="col-10">
-                <label class="form-label">@user.FirstName</label>
+                <label>@user.FirstName</label>
             </div>
         </div>
 
@@ -31,7 +31,7 @@
                 <label class="form-label" asp-for="@user.LastName"></label>
             </div>
             <div class="col-10">
-                <label class="form-label">@user.LastName</label>
+                <label>@user.LastName</label>
             </div>
         </div>
 
@@ -40,7 +40,7 @@
                 <label class="form-label" asp-for="@user.Email"></label>
             </div>
             <div class="col-10">
-                <label class="form-label">@user.Email</label>
+                <label>@user.Email</label>
             </div>
         </div>
 
@@ -49,7 +49,7 @@
                 <label class="form-label" asp-for="@user.Telephone"></label>
             </div>
             <div class="col-10">
-                <label class="form-label">@user.Telephone</label>
+                <label>@user.Telephone</label>
             </div>
         </div>
     </div>
