+++ MedicalSuiteWeb/wwwroot/js/site.js
@@ -2,3 +2,16 @@
 // for details on configuring this project to bundle and minify static web assets.
 
 // Write your JavaScript code.
+function togglePasswordVisibility() {
+    var passwordInput = document.getElementById("password");
+    var eyeIcon = document.getElementById("eyeIcon");
+
+    if (passwordInput.type === "password") {
+        passwordInput.type = "text";
+        eyeIcon.className = "bi bi-eye-slash";
+    } else {
+        passwordInput.type = "password";
+        eyeIcon.className = "bi bi-eye";
+    }
+
+}
